refactor: organisation categories modals refactor: reduce duplicated code
